RENDERER: Add option to filter out collinear vertices.
This is a temporary workaround primarily targeted at macOS arm64 on which
the GPU has a high risk of clipping nearby triangles if zero sized triangles
are fed to the GPU. At best this bug introduces rendering glitches from looking
into the void, at worst it allows the player to see enemies behind walls in matches.

The workaround is force-enabled on macOS arm64. But there have been mysterious
rendering issues with missing surfaces in the past, so other targets may opt-in
to enable it and see if it helps.